feat: :sparkles: makes components are client-only rendering
Fixes #28
